- tratamento de erro variado, devido a ao ter padrão no retorno do erro da API, a rotina tem de satrar vários padrões de retorno de erro.

+++ vc-module-MelhorEnvio/vc_module_MelhorEnvio.Core/Models/CartOut.cs
@@ -125,7 +125,7 @@ namespace vc_module_MelhorEnvio.Core.Models
         public string Format { get; set; }
 
         [JsonProperty("billed_weight")]
-        public int BilledWeight { get; set; }
+        public decimal BilledWeight { get; set; }
 
         [JsonProperty("receipt")]
         public bool Receipt { get; set; }
